trying to fix scope...
