+++ packages/livedoc-mocha/_src/app/Dummy.ts
@@ -1,0 +1,3 @@
+export class Dummy {
+
+}
