--- packages/livedoc-mocha/_src/app/FeatureContext.ts
@@ -1,6 +1,0 @@
-export class FeatureContext {
-    filename: string;
-    title: string;
-    description: string;
-    tags: string[];
-}
