+++ packages/livedoc-mocha/_src/app/Row.ts
@@ -1,3 +1,3 @@
-export interface Row {
+declare interface Row {
     [prop: string]: any;
 }
