--- packages/livedoc-mocha/_src/app/ScenarioContext.ts
@@ -1,10 +1,0 @@
-
-import { StepContext } from "./index";
-
-export class ScenarioContext {
-    title: string;
-    description: string;
-    given: StepContext;
-    and: StepContext[] = [];
-    tags: string[];
-}
