--- packages/livedoc-mocha/_src/app/index.ts
@@ -1,6 +1,0 @@
-export { Row } from './Row';
-export { StepContext } from './StepContext';
-export { ScenarioOutlineContext } from './ScenarioOutlineContext';
-export { ScenarioContext } from './ScenarioContext';
-export { FeatureContext } from './FeatureContext';
-export { BackgroundContext } from "./BackgroundContext";
