+++ packages/livedoc-mocha/_src/test/Background.Spec.ts
@@ -1,6 +1,4 @@
-///<reference path="../app/index.ts" />
 import * as Utils from "./Utils";
-import "../app/index";
 
 require('chai').should();
